pyproject.toml
[project]
name = "scalar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "typer>=0.21.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
